Major refactoring to separate concerns
This commit moves the calculation helpers into an `ogen` subpackage.
Doing so significantly reduces the size of a resulting binary (saved
3MB for me on linux-amd64).
