+++ optimus_test.go
@@ -2,128 +2,44 @@
 
 package optimus
 
-import (
-	"archive/zip"
-	"bytes"
-	"crypto/rand"
-	"fmt"
-	"io/ioutil"
-	// "log"
-	"math/big"
-	"strings"
-	"testing"
-	"unsafe"
-)
+import
 
-// Obtains a prime number from the internet, calculates the mod inverse of it and
-// calculates a random number. It then checks if the process worked BUT does not
-// test if the number obtained is actually Prime.
-func TestGenerateSeed(t *testing.T) {
+// "log"
 
-	for i := 0; i < 3; i++ { //How many times we want to run GenerateSeed()
-		o, err, f := GenerateSeed(nil)
-		if err != nil {
-			t.Errorf("Try %d - Failed", i)
-		}
+"testing"
 
-		//Check if prime is contained in zipped text file
-		baseURL := "http://primes.utm.edu/lists/small/millions/primes%d.zip"
-		finalUrl := fmt.Sprintf(baseURL, f)
+func TestOptimus(t *testing.T) {
+	var encodeTests = []struct {
+		prime  uint64
+		mod    uint64
+		rand   uint64
+		input  uint64
+		encode uint64
+	}{
+		{1580030173, 59260789, 1163945558, 15, 1103647397},
+		{1580030173, 59260789, 1163945558, 99999, 1458223381},
 
-		resp, err := client(nil).Get(finalUrl)
-		if err != nil {
-			t.Errorf("Try %d - Failed", i)
-			continue
-		}
-		defer resp.Body.Close()
+		{2123809381, 1885413229, 146808189, 15, 1645575830},
+		{2123809381, 1885413229, 146808189, 99999, 1124632518},
 
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			t.Errorf("Try %d - Failed", i)
-			continue
-		}
+		{3, 715827883, 1234567890, 15, 1234567935},
+		{3, 715827883, 1234567890, 99999, 1234342159},
 
-		r, err := zip.NewReader(bytes.NewReader(body), resp.ContentLength)
-		if err != nil {
-			t.Errorf("Try %d - Failed", i)
-			continue
-		}
-
-		zippedFile := r.File[0]
-
-		src, err := zippedFile.Open() //src contains ReaderCloser
-		if err != nil {
-			t.Errorf("Try %d - Failed", i)
-			continue
-		}
-		// defer src.Close()
-
-		//Create a Byte Slice
-		buf := new(bytes.Buffer)
-		buf.ReadFrom(src)
-		b := buf.Bytes()
-		stringContents := *(*string)(unsafe.Pointer(&b))
-
-		subString := fmt.Sprintf(" %d ", o.Prime())
-		if !strings.Contains(stringContents, subString) {
-			src.Close()
-			t.Errorf("Try %d - Failed - Obtained Prime:%d is not a Prime", i, o.Prime())
-			continue
-		}
-
-		src.Close()
-
-		//Check if ModInverse is correct
-
-		// if o.Prime() != ModInverse(o.ModInverse()) {
-		// 	src.Close()
-		// 	t.Errorf("Try %d - Failed - ModInverse(%d) of %d is not correct", i, o.ModInverse, o.Prime)
-		// 	continue
-		// }
-
+		{837350711, 1701236871, 1234567890, 1024, 1782767314},
+		{837350711, 1701236871, 1234567890, 1025, 472634341},
 	}
-}
 
-// Tests if the encoding process correctly decodes the id back to the original.
-func TestEncoding(t *testing.T) {
-	for i := 0; i < 15; i++ { //How many times we want to run GenerateSeed()
-		o, _, _ := GenerateSeed(nil)
+	for i, tt := range encodeTests {
+		o := New(tt.prime, tt.mod, tt.rand)
 
-		c := 10
-		h := 100 //How many random numbers to select in between 0-c and (MAX_INT-c) - MAX-INT
-
-		var y []uint64 //Stores all the values we want to run encoding tests on
-
-		for t := 0; t < c; t++ {
-			y = append(y, uint64(t))
+		enc := o.Encode(tt.input)
+		if enc != tt.encode {
+			t.Errorf("[%d] encode failed: expected %d, got %d", i, tt.encode, enc)
 		}
 
-		//Generate Random numbers
-		for t := 0; t < h; t++ {
-			upper := *big.NewInt(int64(MAX_INT - 2*c))
-			rand, _ := rand.Int(rand.Reader, &upper)
-			randomNumber := rand.Uint64() + uint64(c)
-
-			y = append(y, randomNumber)
+		dec := o.Decode(enc)
+		if dec != tt.input {
+			t.Errorf("[%d] decode failed: expected %d, got %d", i, tt.input, dec)
 		}
-
-		for t := MAX_INT; t >= MAX_INT-c; t-- {
-			y = append(y, uint64(t))
-		}
-
-		t.Logf("Prime: %d ModInverse: %d Random: %d", o.Prime(), o.ModInverse(), o.Random())
-		for _, value := range y {
-			orig := value
-			hashed := o.Encode(value)
-			unhashed := o.Decode(hashed)
-
-			if orig != unhashed {
-				t.Errorf("%d: %d -> %d - FAILED", orig, hashed, unhashed)
-			} else {
-				t.Logf("%d: %d -> %d - PASSED", orig, hashed, unhashed)
-				// log.Printf("%d: %d -> %d - PASSED", orig, hashed, unhashed)
-			}
-		}
-
 	}
 }
